+++ services/horizon/internal/db2/core/main.go
@@ -22,6 +22,7 @@ type Account struct {
 	HomeDomain         null.String
 	Thresholds         xdr.Thresholds
 	Flags              xdr.AccountFlags
+	LastModified       uint32
 	BuyingLiabilities  xdr.Int64 `db:"buyingliabilities"`
 	SellingLiabilities xdr.Int64 `db:"sellingliabilities"`
 }
@@ -129,6 +130,7 @@ type Trustline struct {
 	Tlimit             xdr.Int64
 	Balance            xdr.Int64
 	Flags              int32
+	LastModified       uint32
 	BuyingLiabilities  xdr.Int64 `db:"buyingliabilities"`
 	SellingLiabilities xdr.Int64 `db:"sellingliabilities"`
 }