+++ services/horizon/internal/resourceadapter/account.go
@@ -27,6 +27,7 @@ func PopulateAccount(
 	dest.SubentryCount = ca.Numsubentries
 	dest.InflationDestination = ca.Inflationdest.String
 	dest.HomeDomain = ca.HomeDomain.String
+	dest.LastModifiedLedger = ca.LastModified
 
 	PopulateAccountFlags(&dest.Flags, ca)
 	PopulateAccountThresholds(&dest.Thresholds, ca)